requirements.txt
streamlit>=1.42.0
altair>=5.0.0
gspread
google-auth
pandas
plotly
streamlit-autorefresh